phy: marvell10g: Update advertised modes at init
When initializing the PHY configuration with .config_init, the
marvell10g driver dynamically builds a list of the supported
link modes.

The advertised modes have to be updated according to this
new list, since some of them might be missing from the inital list
built from the driver's .features field.

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>