phy: marvell10g: Add support for 2.5GBaseT and 5GBaseT modes
The Marvell Alaska family of PHYs supports 2.5GBaseT and 5GBaseT modes,
as defined in the 802.3bz and the NGBaseT alliance. Some of the PHYs in
this family don't advertise that support in the PMA Speed ability
register as they should.

This was tested on a 88E3310 PHY on a MacchiatoBin platform, and a 88E2010
PHY (similar to the 33xx family, without the fibre interfaces).

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>