net: mvneta: Remove per-cpu queue mapping for Armada 3700
According to Errata #23 "The per-CPU GbE interrupt is limited to Core
0", we can't use the per-cpu interrupt mechanism on the Armada 3700
familly.

This is correctly checked for RSS configuration, but the initial queue
mapping is still done by having the queues spread across all the CPUs in
the system, both in the init path and in the cpu_hotplug path.

Add checks to make sure we pin all queues to CPU 0.

This significantly improves performances, since we now allocate the full
resources of the controller to the only CPU being able to use it. This
show especially when using the MAC on a 2.5Gbps link :

Before this commit :

  TX path (server on DUT) : iperf3 -R -c 192.168.42.2

  RX path (server on DUT) : iperf3 -c 192.168.42.2

[ ID] Interval           Transfer     Bitrate         Retr
[  5]   0.00-10.00  sec  1.63 GBytes  1.40 Gbits/sec  100             sender
[  6]   0.00-10.00  sec  1.63 GBytes  1.40 Gbits/sec                  receiver

After this commit :

  TX path (server on DUT) : iperf3 -R -c 192.168.42.2

[ ID] Interval           Transfer     Bitrate         Retr
[  5]   0.00-10.00  sec  2.74 GBytes  2.35 Gbits/sec    0             sender
[  5]   0.00-10.00  sec  2.74 GBytes  2.35 Gbits/sec                  receiver

  RX path (server on DUT) : iperf3 -c 192.168.42.2

[ ID] Interval           Transfer     Bitrate         Retr
[  5]   0.00-10.00  sec  1.62 GBytes  1.39 Gbits/sec   77             sender
[  5]   0.00-10.00  sec  1.62 GBytes  1.39 Gbits/sec                  receiver

Note that on the RX path, we start seeing overrun errors at that speed

In TX path : From 1.18 Gbps to 2.35 Gbps

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>