net: mvneta: Implement per-queue shaping offload
This commit implements traffic shaping on a per-queue basis, using the
Traffig Shaper Version 3 present in mvneta.

It uses a Token-Bucket algorithm, with a shared clock but per-queue
token size, refill rate and refill value.

The hardcoded token refill rate allows for a resolution of 10Kbps and
can be used to express rate-limit from 10Kbps per queue up to 5Gbps.

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>