net: stmmac: don't rely on lynx_pcs presence to check for a PHY
When initializing attached PHYs, there are some cases where we don't
expect any PHY to be connected. The logic uses conditions based on
various local PCS configuration, but also calls-in
phylink_expects_phy(), which is enough to ensure we don't try to
initialize a PHY when using a Lynx PCS, as long as we have the
phy_interface set to a 802.3z mode and usng inband negociation.

Drop the lynx check, making the stmmac generic code more
pcs_lynx-agnostic.

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>