net: stmmac: dwmac_socfpga: initialize the PCS before registering the netdev
As reported by Russell[1], there are possible race conditions in
dwmac_socfpga due to initializing the PCS after registering the netdev.

This was due to the fact that the lynx pcs reference was stored in the
stmmac_priv data, which was created in the generic stmmac_drv_probe()
function, that registers netdev.

Now that pcs_lynx no longer lives within the generic stmmac_priv data,
but rather in the local socfpga_dwmax bsp private data, we can safely
initialize the PCS before calling into the driver probe function.

[1]: https://lore.kernel.org/netdev/ZH9XK5yEGyoDMIs%2F@shell.armlinux.org.uk/

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>