net: phy: split phy_attach_direct between setup and attachment
phy_attach_direct() is used to associate a phy_device to a net_device.
It contains phy initialisation code as well as the logic to link the
net_device and phy_device together.

Split the phy setup phase into a dedicated function, preparing the
support for stacked PHYs

TODO : split-out no_inband

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>