net: ethtool: Add mechanism to prepare notifications
Some ethtool notifications may need some request context to build the
notification message. As we don't have .parse_request() involved in notifs,
add a .prepare_ntf() callback that has the same role of populating the
req_info context.

Because we don't have any user request we can base ourselves on to build
req_info, we rely instead on the fact that ethnl_notify() is passed a
const void *data pointer.

This is passed to the new .prepare_ntf() callback, which can use it
however it wants to build a meaningful context.

The PORT Notification is here used as an example.

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>