net: ethtool: Convert PHY dump into Topology dump
With phy_link_topology, we can query about the topology of what sits
after a MAC controller. There can be multiple PHYs, as well as multiple
ports :                  +-------------+
                      +--| Port (RJ45) |
 +--------+  +-----+  |  +-------------+
 | netdev |--| PHY |--+
 +--------+  +-----+  | +-----+  +-----+ +--------------+
                      +-| SFP |--| PHY |-| Port (Fiber) |
		        +-----+  +-----+ +--------------+

Some of the netlink commands that already exists can be adjusted to give
specific information about one of these nodes. For example, what port is
currently in use ? What does it support ? How can I change which one is
used ?

Port and PHYs have an index, maintaind by phy_link_topology, to report
them to userspace and allow it to interact with it.

Modify the existing helper to dump per-phy info, allowing per-port
dumps.

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>